requirements.txt
httpx
requests
python-dotenv
beautifulsoup4
langchain
langchain-google-genai
langchain-community
langchain-core
tiktoken
pytest
pillow
streamlit
Markdown
google-genai
pydantic
moviepy==1.0.3
gTTS
pysrt
edge-tts
pydub 
librosa 
numpy 
soundfile
opencv-python
pexels-api